+++ codersmuse/config.py
@@ -4,6 +4,7 @@ PLUGIN_BEHAVORIAL_ACTIVE = True
 PLUGIN_EYETRACKING_ACTIVE = True
 PLUGIN_PHYSIO_ACTIVE = True
 PLUGIN_FMRI_ACTIVE = False  # Set to True
+PLUGIN_EEG_ACTIVE = True  # Set to False
 
 EYETRACKING_DRAW_MODE = 'saccades'  # alternatively, use 'gazepath'
 EYETRACKING_LENGTH_TRACE = 150  # in milliseconds