+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Fixed issues with newer library versions
 - Fixed issue with non-existing temp folders
+- Prepare EEG viewer
 
 ### [0.1.1] - 2020-04-09
 - Fixed some start issues with default settings and an initialized project